[Refactor][Kernel] Flatten single-file kernel directories
Collapse deltanet_recurrence/, gla_recurrence/, gated_deltanet_recurrence/,
fft/, grouped_gemm/ to flat modules. Merge gemm/gemm.py + gemm/gemv.py
into a single gemm.py. Fix relative imports and cross-references.

Part 6 of #924.
