--- tileops/kernels/deltanet_recurrence/__init__.py
@@ -1,6 +1,0 @@
-from .deltanet_decode import DeltaNetDecodeFP32Kernel, DeltaNetDecodeKernel
-
-__all__ = [
-    "DeltaNetDecodeFP32Kernel",
-    "DeltaNetDecodeKernel",
-]
