--- tileops/kernels/fft/__init__.py
@@ -1,3 +1,0 @@
-from .fft_c2c import FFTC2CKernel
-
-__all__ = ["FFTC2CKernel"]
