--- tileops/kernels/gated_deltanet_recurrence/__init__.py
@@ -1,6 +1,0 @@
-from .gated_deltanet_decode import GatedDeltaNetDecodeFP32Kernel, GatedDeltaNetDecodeKernel
-
-__all__ = [
-    "GatedDeltaNetDecodeFP32Kernel",
-    "GatedDeltaNetDecodeKernel",
-]
