--- tileops/kernels/gemm/__init__.py
@@ -1,4 +1,0 @@
-from .gemm import GemmKernel
-from .gemv import GemvKernel
-
-__all__ = ["GemmKernel", "GemvKernel"]
