--- tileops/kernels/gla_recurrence/__init__.py
@@ -1,6 +1,0 @@
-from .gla_decode import GLADecodeFP32Kernel, GLADecodeKernel
-
-__all__ = [
-    "GLADecodeFP32Kernel",
-    "GLADecodeKernel",
-]
