--- tileops/kernels/grouped_gemm/__init__.py
@@ -1,5 +1,0 @@
-from .grouped_gemm import GroupedGemmKernel
-
-__all__ = [
-    "GroupedGemmKernel",
-]
